correction to download check

+++ examples/cluster_blueprint/scripts/codes/build.sh
@@ -18,7 +18,7 @@ go install src/${PACKAGEPATH}/cfy-go/cfy-go.go
 ctx logger info "Attempting to download cfy-kubernetes from CFY Manager"
 KUBERNETES_BINARY=$(ctx download-resource resources/cfy-kubernetes)
 
-if [[ $? == 0 ]]; then
+if [[ $? == 0 ]] && [[ -e "$KUBERNETES_BINARY" ]]; then
     ctx logger info "Downloaded provided cfy-kubernetes"
     cp $KUBERNETES_BINARY /opt/cloudify-kubernetes-provider/bin/cfy-kubernetes
 else
